+++ cli/golang_internal_testsuite/testsuite_impl/internal_testsuite.go
@@ -6,18 +6,18 @@
 package testsuite_impl
 
 import (
-	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/advanced_network_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/basic_datastore_and_api_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/basic_datastore_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/bulk_command_execution_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/exec_command_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/files_artifact_mounting_test"
-	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/lambda_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/local_static_file_test"
+	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/module_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/network_partition_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/test_internal_state_persistence_test"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/testsuite_impl/wait_for_endpoint_availability_test"
+	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
 )
 
 
@@ -50,9 +50,9 @@ func (suite InternalTestsuite) GetTests() map[string]testsuite.Test {
 		"waitForEndpointAvailabilityTest": wait_for_endpoint_availability_test.NewWaitForEndpointAvailabilityTest(
 			suite.datastoreServiceImage,
 		),
-		"localStaticFileTest": local_static_file_test.LocalStaticFileTest{},
-		"bulkCommandExecutionTest": bulk_command_execution_test.NewBulkCommandExecutionTest(suite.datastoreServiceImage),
-		"lambdaTest": lambda_test.LambdaTest{},
+		"localStaticFileTest":              local_static_file_test.LocalStaticFileTest{},
+		"bulkCommandExecutionTest":         bulk_command_execution_test.NewBulkCommandExecutionTest(suite.datastoreServiceImage),
+		"moduleTest":                       module_test.ModuleTest{},
 		"testInternalStatePersistenceTest": test_internal_state_persistence_test.NewTestInternalStatePersistenceTest(),
 	}
 	return tests