+++ cli/docs/changelog.md
@@ -1,4 +1,13 @@
 # TBD
+### Changes
+* Switched all references to "Lambda" to "module"
+
+### Fixes
+* `ModuleTest` (renamed from `LambdaTest`) now uses the ports returned by the Datastore Army module
+* Fixed bug in CI where `pipefail` wasn't set which would result in the testsuite-running step passing when it shouldn't
+
+### Breaking Changes
+* Renamed the `lambda` command to `module`
 
 # 0.3.4
 ### Fixes